// src/app/modules/Manager/manager.interface.ts
import { Types } from "mongoose";

export interface TManager {
    managerId?: string;
    user: Types.ObjectId;
    name: {
        firstName: string;
        lastName: string;
    };
    email: string;
    password: string;
    phoneNumber: string;
    profilePhotoUrl?: string;
    isDeleted: boolean;
}